requirements.txt
langchain===0.0.239
openai
vertexai
google-cloud-aiplatform
qdrant-client
tiktoken
docx2txt
pypdf
cohere
underthesea
chroma
chromadb
pydantic==1.10.7
google-cloud-translate
streamlit==1.24.0
pdfplumber
python-docx
seaborn
matplotlib
plotly
anthropic==0.3.4
Office365-REST-Python-Client==2.3.16
msal-streamlit-authentication==1.0.9
httpx-oauth
openpyxl==3.1.2
pyxlsb==1.0.10
pysqlite3
pysqlite3-binary